+++ types/query/collections_pagination.go
@@ -3,7 +3,9 @@ package query
 import (
 	"context"
 	"errors"
-	"fmt"
+
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
 
 	"cosmossdk.io/collections"
 	collcodec "cosmossdk.io/collections/codec"
@@ -79,7 +81,7 @@ func CollectionFilteredPaginate[K, V any, C Collection[K, V], T any](
 	reverse := pageReq.Reverse
 
 	if offset > 0 && key != nil {
-		return nil, nil, fmt.Errorf("invalid request, either offset or key is expected, got both")
+		return nil, nil, status.Error(codes.InvalidArgument, "invalid request, either offset or key is expected, got both")
 	}
 
 	opt := new(CollectionsPaginateOptions[K])